stub CommentButtons component
